PRC-101: extracted out the ProxyClient into it's own file
